Improve handling of over sized initial decks
v2.4.3 introduced a "shrink" function to fix the issue that climb would
sometimes result in a >10 card deck when given a >10 card initial deck.

Move the call to shrink until AFTER the call to claim_cards has happened.
This allows user to pass an over sized deck on purpose to a climb and get
the best 10 card deck without having to muck around with setting up
owned cards.